+++ saleor/graphql/checkout/tests/mutations/test_order_create_from_checkout.py
@@ -12,7 +12,13 @@ from .....checkout import calculations
 from .....checkout.error_codes import OrderCreateFromCheckoutErrorCode
 from .....checkout.fetch import fetch_checkout_info, fetch_checkout_lines
 from .....checkout.models import Checkout, CheckoutLine
-from .....core.taxes import TaxError, zero_money, zero_taxed_money
+from .....core.taxes import (
+    TaxDataError,
+    TaxDataErrorMessage,
+    TaxError,
+    zero_money,
+    zero_taxed_money,
+)
 from .....discount import DiscountType, DiscountValueType, RewardValueType
 from .....discount.models import CheckoutLineDiscount
 from .....giftcard import GiftCardEvents
@@ -2501,3 +2507,32 @@ def test_order_from_draft_create_0_total_value_from_giftcard(
     assert not Checkout.objects.filter(pk=checkout.pk).exists(), (
         "Checkout should have been deleted"
     )
+
+
+@patch("saleor.checkout.calculations.validate_tax_data")
+def test_order_from_checkout_tax_error(
+    mock_validate_tax_data,
+    app_api_client,
+    permission_handle_checkouts,
+    checkout_with_items_and_shipping,
+    caplog,
+):
+    # given
+    mock_validate_tax_data.side_effect = TaxDataError(TaxDataErrorMessage.EMPTY)
+    checkout = checkout_with_items_and_shipping
+    variables = {"id": graphene.Node.to_global_id("Checkout", checkout.pk)}
+
+    # when
+    response = app_api_client.post_graphql(
+        MUTATION_ORDER_CREATE_FROM_CHECKOUT,
+        variables,
+        permissions=[permission_handle_checkouts],
+    )
+    content = get_graphql_content(response)
+
+    # then
+    data = content["data"]["orderCreateFromCheckout"]
+    assert data["errors"][0]["code"] == OrderCreateFromCheckoutErrorCode.TAX_ERROR.name
+    assert data["errors"][0]["field"] is None
+    assert not Order.objects.exists()
+    assert "Tax app error for checkout" in caplog.text