+++ saleor/graphql/payment/tests/mutations/test_transaction_update.py
@@ -243,6 +243,46 @@ def test_transaction_update_metadata_by_app(
     assert transaction_item_created_by_app.metadata == {meta_key: meta_value}
 
 
+def test_transaction_update_metadata_by_app_metadata_extended(
+    transaction_item_created_by_app, permission_manage_payments, app_api_client
+):
+    # given
+    transaction = transaction_item_created_by_app
+    old_key = "old-key"
+    old_value = "old-value"
+    transaction.metadata = {old_key: old_value}
+    transaction.save(update_fields=["metadata"])
+
+    meta_key = "key-name"
+    meta_value = "key_value"
+    variables = {
+        "id": graphene.Node.to_global_id("TransactionItem", transaction.token),
+        "transaction": {
+            "metadata": [{"key": meta_key, "value": meta_value}],
+        },
+    }
+
+    # when
+    response = app_api_client.post_graphql(
+        MUTATION_TRANSACTION_UPDATE, variables, permissions=[permission_manage_payments]
+    )
+
+    # then
+    transaction.refresh_from_db()
+    content = get_graphql_content(response)
+    data = content["data"]["transactionUpdate"]["transaction"]
+    assert len(data["metadata"]) == 2
+    assert {metadata["key"] for metadata in data["metadata"]} == {old_key, meta_key}
+    assert {metadata["value"] for metadata in data["metadata"]} == {
+        old_value,
+        meta_value,
+    }
+    assert transaction_item_created_by_app.metadata == {
+        old_key: old_value,
+        meta_key: meta_value,
+    }
+
+
 def test_transaction_update_metadata_by_app_null_value(
     transaction_item_created_by_app, permission_manage_payments, app_api_client
 ):
@@ -325,6 +365,50 @@ def test_transaction_update_private_metadata_by_app(
     assert data["privateMetadata"][0]["key"] == meta_key
     assert data["privateMetadata"][0]["value"] == meta_value
     assert transaction_item_created_by_app.private_metadata == {meta_key: meta_value}
+
+
+def test_transaction_update_private_metadata_by_app_extend_metadata(
+    transaction_item_created_by_app, permission_manage_payments, app_api_client
+):
+    # given
+    transaction = transaction_item_created_by_app
+
+    old_key = "old-key"
+    old_value = "old-value"
+    transaction.private_metadata = {old_key: old_value}
+    transaction.save(update_fields=["private_metadata"])
+
+    meta_key = "key-name"
+    meta_value = "key_value"
+    variables = {
+        "id": graphene.Node.to_global_id("TransactionItem", transaction.token),
+        "transaction": {
+            "privateMetadata": [{"key": meta_key, "value": meta_value}],
+        },
+    }
+
+    # when
+    response = app_api_client.post_graphql(
+        MUTATION_TRANSACTION_UPDATE, variables, permissions=[permission_manage_payments]
+    )
+
+    # then
+    transaction.refresh_from_db()
+    content = get_graphql_content(response)
+    data = content["data"]["transactionUpdate"]["transaction"]
+    assert len(data["privateMetadata"]) == 2
+    assert {metadata["key"] for metadata in data["privateMetadata"]} == {
+        old_key,
+        meta_key,
+    }
+    assert {metadata["value"] for metadata in data["privateMetadata"]} == {
+        old_value,
+        meta_value,
+    }
+    assert transaction_item_created_by_app.private_metadata == {
+        old_key: old_value,
+        meta_key: meta_value,
+    }
 
 
 def test_transaction_update_private_metadata_by_app_null_value(