+++ saleor/tests/e2e/checkout/taxes/test_checkout_complete_return_tax_error.py
@@ -133,4 +133,7 @@ def test_checkout_calculate_return_tax_error_when_app_not_respond_CORE_2013(
     )
     assert order_data["errors"] is not None
     assert order_data["errors"][0]["code"] == "TAX_ERROR"
-    assert order_data["errors"][0]["message"] == "Configured Tax App didn't responded."
+    assert (
+        order_data["errors"][0]["message"]
+        == "Configured Tax App returned invalid response."
+    )